src/test/java/com/aazizova/springboottesttask/controller/ProductControllerTest.java
package com.aazizova.springboottesttask.controller;

import com.aazizova.springboottesttask.model.entity.Product;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import java.util.ArrayList;
import java.util.List;

import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
import static org.mockito.Mockito.mock;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@SpringBootTest
@AutoConfigureMockMvc
public class ProductControllerTest {
    @Autowired
    private ProductController productController;

    private List<Product> products;

    @Autowired
    private MockMvc mockMVC;

    @Before
    public void setUp() throws Exception {
        products = mock(ArrayList.class);
    }

    @Test
    public void contextLoads() throws Exception {
        assertThat(productController).isNotNull();
    }

    @Test
    public void authIsRequiredTest() throws Exception {
        mockMVC.perform(get("/api/products"))
                .andDo(print())
                .andExpect(status().isUnauthorized());
    }

    @Test
    @WithMockUser(roles = "ROLE_ADMIN")
    public void getAllProductsTest() throws Exception {
        mockMVC.perform(get("/api/products"))
                .andDo(print())
                .andExpect(status().isOk());
    }

    @Test
    @WithMockUser
    public void exportProductsTest() throws Exception {
        this.mockMVC.perform(get("/api/products/export")
                .contentType(MediaType.APPLICATION_JSON)
                .content(toJson(products)))
                .andDo(print())
                .andExpect(status().isOk());
    }

    private static String toJson(final Object obj) {
        try {
            return new ObjectMapper().writeValueAsString(obj);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }
}